Extract cropTable() utility function.
